+++ src/routes/(v2)/v2/issue-360/target/+page.svelte
@@ -1,0 +1,2 @@
+<p>First redirect target</p>
+<a href="/">Go home</a>