+++ src/routes/(v2)/v2/issue-360/target/+page.ts
@@ -1,0 +1,5 @@
+import { redirect } from '@sveltejs/kit';
+
+export function load() {
+	redirect(303, '/v2/issue-360/second-target');
+}